+++ CHANGELOG.md
@@ -14,3 +14,7 @@ Changed for individual apps are listed in `apps/appname/ChangeLog`
 * Added espruinotools.js for pretokenisation
 * Included image and compression tools in repo
 * Added better upload of large files (incl. compression)
+* URL fetch is now async
+* Adding '#search' after the URL (when not the name of a 'filter' chip) will set up search for that term
+* If `bin/pre-publish.sh` has been run and recent.csv created, add 'Sort By' chip
+* New 'espruinotools' which fixes pretokenise issue when ID follows ID (fix #416)